upgrade api, create payment method, billDetail

+++ f4education/backend/src/main/java/com/f4education/springjwt/payload/response/BillResponseDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class BillResponseDTO {
 	private Integer billId;
 	private Date createDate;
-	private Float totalPrice;
+	private Double totalPrice;
 	private String status;
 	private String notes;
 	private String paymentMethod;
